# [DataGridPremium] Improve aggregation performance for multiple columns