@cherniavskii 2025-01-08
> Unrelated optimization but we could speed this further by replacing the object params with regular params.

Thanks for the suggestion, I'll give it a try!