@cherniavskii 2025-01-08
Changing to regular parameters didn't bring significant improvements (547ms median compared to 548ms before the change).
I'll keep it anyway. I also spotted a couple of potential small optimizations around that part of the code.